"check_style_score -> 3 of 4 passed. style score: 0/1 "

+++ cspp2-assignments/m17/Shoppingcart (1)/Shoppingcart/Solution.java
@@ -139,7 +139,7 @@ class ShoppingCart {
     /**
      * { number num1}.
      */
-    private static final int NUM1 = 10;
+    private static final float NUM1 = 10f;
     /**
      * { string arr}.
      */
